requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
dlib==19.22.1
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==3.3
numpy==1.21.4
opencv-python==4.5.4.60
Pillow==8.4.0
python-dotenv==0.19.2
pyttsx3==2.90
pyzmq==22.3.0
requests==2.26.0
six==1.16.0
torch==1.10.1
torch-vision==0.1.6.dev0
typing_extensions==4.0.1
urllib3==1.26.7
